Log message: <LOG_BODY> - 80-bit long double support for NEAT ALU <CODE_REVIEW> - http://cvcc-ubu-01.iil.intel.com/cr/r/850/ <PERF_IMPLICATIONS> - None. <TEST> - Pre-commit http://cvcc-fc14-01.iil.intel.com:8080/job/PCompiler_yveselov/, local unittest.

+++ opencl-intel/backend/validations/Comparator/ComparisonResults.h
@@ -30,7 +30,7 @@ namespace Validation
     {
         IMemoryObjectDescPtr pDesc;
         uint64_t numMismatches;
-        long double maxDiff;
+        double maxDiff;
         float maxInterval;
 
         CompStatistics() : pDesc(NULL), numMismatches(0), maxDiff(0.0), maxInterval(0.0)
